+++ install/install.sql
@@ -8,6 +8,7 @@ CREATE TABLE IF NOT EXISTS `glpi_plugin_maintenancecosts_materials` (
    `unit` varchar(32) NOT NULL DEFAULT '',
    `category` varchar(255) NOT NULL DEFAULT '',
    `is_active` tinyint NOT NULL DEFAULT '1',
+   `has_current_quote` tinyint NOT NULL DEFAULT '0',
    `date_creation` timestamp NULL DEFAULT NULL,
    `date_mod` timestamp NULL DEFAULT NULL,
    PRIMARY KEY (`id`),
@@ -15,6 +16,7 @@ CREATE TABLE IF NOT EXISTS `glpi_plugin_maintenancecosts_materials` (
    KEY `idx_entity` (`entities_id`),
    KEY `idx_recursive` (`is_recursive`),
    KEY `idx_active` (`is_active`),
+   KEY `idx_current_quote` (`has_current_quote`),
    KEY `idx_name` (`name`),
    KEY `idx_category` (`category`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci ROW_FORMAT=DYNAMIC;